Made "Rating" property a floating point number for more accuracy when rating a song and rounded it to the second decimal place when listing.

+++ SongCatalog/Models/Song.cs
@@ -6,7 +6,7 @@ namespace SongCatalog.Models
 {
     public class Song
     {
-        public Song(string title, string artistName, int rating)
+        public Song(string title, string artistName, float rating)
         {
             Title = title;
             ArtistName = artistName;
@@ -18,6 +18,6 @@ namespace SongCatalog.Models
         public string ArtistName { get; set; } = null!;
 
         [Range(RatingMinValue, RatingMaxValue, ErrorMessage = InvalidRatingMessage)]
-        public int Rating { get; set; }
+        public float Rating { get; set; }
     }
 }
